将gen..._panel_new.py合并入gen..._panel.py,如今是我在集群上之版本,以及修改sh脚本名称去掉了'_new'
