+++ report.md
@@ -1,0 +1,56 @@
+使用新数据生成方法生成了单圈与双圈数据各1万个。双圈数据之压缩方法为对x方向做压缩变换:
+$$x' = (x-2.5) * 0.49 + 2.5; x' = (x+2.5) * 0.49 - 2.5 $$
+目前我之结果:皆为四种核在单圈上做训练(训练集大小9千)并在双圈上做测试(测试集大小1千)
+
+四种核:
+1. sp_laplace: 单层势
+2. dp_laplace: 偶极子势
+3. modified_laplace: 去掉了与法向之点积的二维值偶极子势卷积核
+4. stokes: stokes核
+
+诸参数方面:
+1. 四种核中唯有dp_laplace需要"normal_prod=True"
+2. 每个测试的四种参数设定与记录名称:
+   
+   | grad | geo | lap | name |
+   | :--- | :--- | :--- | :--- |
+   | true | false | false | noact |
+   | true | true | false | noact_geo3wx |
+   | false | true | false | noact_nograd_geo3wx |
+   | false | false | false | noact_nograd |
+
+对于4中参数设定(或称方法)对于不同积分核在一层无激活函数神经网络(即线性回归)中之结果,以下皆只写出在单圈上的训练误差和在双圈上的测试误差,略去百分号。
+
+注:有无"no_grad"是会带来差别的，只是可能比较微小以至于在舍入中忽略了。
+
+1. stokes核
+   
+   | k_max | noact | noact_geo3wx | noact_nograd | noact_nograd_geo3wx |
+   | :--- | :--- | :--- | :--- | :--- |
+   | 8  | 5.60,4.72 | 1.68,2.06 | 5.60,4.73 | 1.71,2.09 |
+   | 16 | 1.94,2.15 | 0.48,0.92 | 1.94,2.15 | 0.48,0.92 |
+   | 32 | 0.53,1.08 | 0.16,0.71 | 0.53,1.08 | 0.15,0.70 |
+
+2. sp_laplace核
+
+    | k_max | noact | noact_geo3wx | noact_nograd | noact_nograd_geo3wx |
+    | :--- | :--- | :--- | :--- | :--- |
+    | 8  | 0.60,0.80 | 0.57,0.78 | 0.60,0.80 | 0.56,0.77 |
+    | 16 | 0.16,0.56 | 0.14,0.56 | 0.16,0.55 | 0.14,0.56 |
+    | 32 | 0.06,0.57 | 0.06,0.56 | 0.06,0.56 | 0.06,0.57 |
+
+3. dp_laplace核
+
+    | k_max | noact | noact_geo3wx | noact_nograd | noact_nograd_geo3wx |
+    | :--- | :--- | :--- | :--- | :--- |
+    | 8  | 14.1,25.7 | 10.6,21.4 | 30.9,43.5 | 10.7,21.1 |
+    | 16 | 6.36,15.1 | 4.24,11.5 | 17.1,32.3 | 4.42,11.8 |
+    | 32 | 2.29,7.39 | 1.45,5.47 | 7.50,17.8 | 1.56,5.68 |
+
+4. modified_dp_laplace核
+
+    | k_max | noact | noact_geo3wx | noact_nograd | noact_nograd_geo3wx |
+    | :--- | :--- | :--- | :--- | :--- |
+    | 8  | 14.1,18.4 | 5.07,9.40 | 15.1,19.1 | 10.6,12.9 |
+    | 16 | 6.50,10.5 | 2.55,5.73 | 6.64,10.6 | 4.93,8.23 |
+    | 32 | 2.44,5.25 | 1.86,4.56 | 2.47,5.26 | 2.00,4.72 |
