+++ scripts/curve/pcno_train_log.sh
